refactors the code into seperate modules.
